changes to auth
