--- src/pages/googlepage.tsx
@@ -1,15 +1,0 @@
-// src/GoogleSignInPage.tsx
-import React from 'react';
-import GoogleSignIn from '../components/googlelog';
-
-const GoogleSignInPage: React.FC = () => {
-  return (
-    <div>
-      <h2>Google Sign-In Page</h2>
-      <GoogleSignIn />
-    </div>
-  );
-};
-
-export default GoogleSignInPage;
-export {};
