--- src/pages/loginpage.tsx
@@ -1,15 +1,0 @@
-// src/LoginPage.tsx
-import React from 'react';
-import Login from '../components/login';
-
-const LoginPage: React.FC = () => {
-  return (
-    <div>
-      <h2>Login Page</h2>
-      <Login />
-    </div>
-  );
-};
-
-export default LoginPage;
-export {};
