--- src/pages/registerpage.tsx
@@ -1,15 +1,0 @@
-// src/RegisterPage.tsx
-import React from 'react';
-import Register from '../components/register';
-
-const RegisterPage: React.FC = () => {
-  return (
-    <div>
-      <h2>Register Page</h2>
-      <Register />
-    </div>
-  );
-};
-
-export default RegisterPage;
-export {};
